requirements.txt
discord.py
beautifulsoup4
aiohttp
pytz
python-dotenv
googletrans==3.1.0a0
requests
flask